requirements.txt
pandas
numpy
schema
pyparsing
biopython
kj
funcy
pyvcf
toolz
sh
mock
